Fix incorrect regexes to preserve doc tags spacing
