+++ packages/better-php-doc-parser/src/PhpDocNode/Doctrine/Class_/AbstractIndexTagValueNode.php
@@ -55,6 +55,7 @@ abstract class AbstractIndexTagValueNode extends AbstractDoctrineTagValueNode im
         if ($originalContent !== null) {
             $this->resolveOriginalContentSpacingAndOrder($originalContent);
         }
+
         $this->tag = $originalTag;
     }
