+++ packages/better-php-doc-parser/tests/PhpDocParser/OrmTagParser/Class_/Fixture/FormattingDoctrineEntity.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\BetterPhpDocParser\Tests\PhpDocParser\OrmTagParser\Class_\Fixture;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Table(
+ *     name="my_entity",
+ *     indexes={
+ *         @ORM\Index(
+ *             name="my_entity_idx", columns={"x", "xx", "xxx", "xxxx"}
+ *         ),
+ *         @ORM\Index(
+ *             name="my_entity_xxx_idx", columns={"xxx"}
+ *         )
+ *     }
+ * )
+ * @ORM\Entity(repositoryClass="App\Repository\MyEntityRepository")
+ */
+final class FormattingDoctrineEntity
+{
+}
